Add a context file to establish the treeData state as a global variable
