--- packages/core/client/src/components/Drawer/sampleData.json
@@ -1,23 +1,0 @@
-[
-  {
-    "id": 1,
-    "parent": 0,
-    "droppable": true,
-    "text": "Documents"
-  },
-  {
-    "id": 3,
-    "parent": 1,
-    "droppable": true,
-    "text": "Uploaded Knowledge"
-  },
-
-
-  {
-    "id": 6,
-    "parent": 0,
-    "droppable": true,
-    "text": "Spells & Prompts"
-  }
-
-]
