+++ packages/core/client/src/contexts/index.ts
@@ -1,1 +1,2 @@
 export { ProjectWindowProvider, useProjectWindow } from './ProjectWindowContext'
+export { TreeDataProvider, useTreeData } from './TreeDataProvider'
